+++ pyobjc-core/Lib/objc/_dyld.py
@@ -13,21 +13,7 @@ __all__ = [
 import os
 
 from objc._framework import infoForFramework
-
-try:
-    from objc._objc import _dyld_shared_cache_contains_path
-except ImportError:
-    _dyld_shared_cache_contains_path = None
-
-
-def dyld_shared_cache_contains_path(p):
-    if _dyld_shared_cache_contains_path is None:
-        return False
-    try:
-        return _dyld_shared_cache_contains_path(p)
-    except NotImplementedError:
-        return False
-
+from objc._objc import _dyld_shared_cache_contains_path
 
 # These are the defaults as per man dyld(1)
 #
@@ -107,7 +93,7 @@ def dyld_framework(filename, framework_name, version=None):
                 yield os.path.join(path, framework_name + ".framework", framework_name)
 
     for f in inject_suffixes(_search()):
-        if dyld_shared_cache_contains_path(f):
+        if _dyld_shared_cache_contains_path(f):
             return f
 
         if os.path.exists(f):
@@ -134,7 +120,7 @@ def dyld_library(filename, libname):
             yield os.path.join(path, libname)
 
     for f in inject_suffixes(_search()):
-        if dyld_shared_cache_contains_path(f):
+        if _dyld_shared_cache_contains_path(f):
             return f
         if os.path.exists(f):
             return f